requirements.txt
streamlit
numpy
pillow
scikit-learn
tensorflow 
keras
